% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetaData.R
\name{getMetaData}
\alias{getMetaData}
\title{Download meta data for xrepository content}
\usage{
getMetaData(identifier)
}
\arguments{
\item{identifer}{The identifier ("kennung")}
}
\value{
The metadata as a list.
}
\description{
Download meta data for xrepository content
}
\examples{
id <- "urn:de:bund:destatis:bevoelkerungsstatistik:schluessel:staatsangehoerigkeit"
getMetaData(id)
}
