% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getValidVersion.R
\name{getValidVersion}
\alias{getValidVersion}
\title{Download meta data for xrepository content}
\usage{
getValidVersion(identifier)
}
\arguments{
\item{identifer}{The identifier ("kennung")}
}
\value{
The identifier of the valid version of the codelist
}
\description{
Download meta data for xrepository content
}
